+++ src/hooks/useCount.tsx
@@ -1,0 +1,25 @@
+import React from "react";
+import { toast } from "react-toastify";
+
+export function useCount(initialValue: number) {
+  const [count, setCount] = React.useState<number>(initialValue);
+
+  const handleIncrementProduct = () => {
+    setCount(count + 1);
+  };
+
+  const handleDecrementProduct = () => {
+    if (count < 2) {
+      return toast.warning("To add to cart you need at least one quantity!");
+    }
+
+    setCount(count - 1);
+  };
+
+  return {
+    count,
+    setCount,
+    handleDecrementProduct,
+    handleIncrementProduct,
+  };
+}
